+++ packages/ast-spec/src/expression/NewExpression/spec.ts
@@ -2,11 +2,11 @@ import type { AST_NODE_TYPES } from '../../ast-node-types';
 import type { BaseNode } from '../../base/BaseNode';
 import type { TSTypeParameterInstantiation } from '../../special/TSTypeParameterInstantiation/spec';
 import type { CallExpressionArgument } from '../../unions/CallExpressionArgument';
-import type { LeftHandSideExpression } from '../../unions/LeftHandSideExpression';
+import type { Expression } from '../../unions/Expression';
 
 export interface NewExpression extends BaseNode {
   type: AST_NODE_TYPES.NewExpression;
-  callee: LeftHandSideExpression;
+  callee: Expression;
   arguments: CallExpressionArgument[];
   typeArguments: TSTypeParameterInstantiation | undefined;