+++ packages/typescript-eslint/src/configs/disable-type-checked.ts
@@ -33,6 +33,7 @@ export default (
     '@typescript-eslint/no-unnecessary-boolean-literal-compare': 'off',
     '@typescript-eslint/no-unnecessary-condition': 'off',
     '@typescript-eslint/no-unnecessary-qualifier': 'off',
+    '@typescript-eslint/no-unnecessary-template-expression': 'off',
     '@typescript-eslint/no-unnecessary-type-arguments': 'off',
     '@typescript-eslint/no-unnecessary-type-assertion': 'off',
     '@typescript-eslint/no-unsafe-argument': 'off',
@@ -68,5 +69,11 @@ export default (
     '@typescript-eslint/unbound-method': 'off',
     '@typescript-eslint/use-unknown-in-catch-callback-variable': 'off',
   },
-  languageOptions: { parserOptions: { project: false, program: null } },
+  languageOptions: {
+    parserOptions: {
+      project: false,
+      program: null,
+      EXPERIMENTAL_useProjectService: false,
+    },
+  },
 });